Improve debug messages
Changed RSYNCFROM_VERSION from "1.21 2022-11-13" to "1.22 2022-11-16".
Changed RSYNCTO_VERSION from "1.21 2022-11-13" to "1.22 2022-11-16".

Improved debug messages.

In rsyncto $DEST_PATH was renamed $DEST_DIR_PATH.

Changed rsyncto last arg to [user@]host[:dir] to help indicate that
the optional path on host is a directory into which the basename
of the from is sent.